requirements.txt
fastapi
uvicorn
reportlab
openpyxl
pandas
psycopg2-binary
sqlalchemy
python-dotenv
